fix framesize division

+++ src/r3d_sprite.c
@@ -10,8 +10,8 @@ R3D_Sprite R3D_LoadSprite(Texture2D texture, int xFrameCount, int yFrameCount)
     sprite.material.maps[MATERIAL_MAP_ALBEDO].texture = texture;
     sprite.material.maps[MATERIAL_MAP_OCCLUSION].value = 1.0f;
 
-    sprite.frameSize.x = texture.width / xFrameCount;
-    sprite.frameSize.y = texture.height / yFrameCount;
+    sprite.frameSize.x = (float)texture.width / xFrameCount;
+    sprite.frameSize.y = (float)texture.height / yFrameCount;
 
     sprite.xFrameCount = xFrameCount;
     sprite.yFrameCount = yFrameCount;
